chore: Extract exception details to doctest/parts/exceptions.h
